+++ clang-tools-extra/clang-tidy/readability/ReadabilityTidyModule.cpp
@@ -53,6 +53,7 @@
 #include "UniqueptrDeleteReleaseCheck.h"
 #include "UppercaseLiteralSuffixCheck.h"
 #include "UseAnyOfAllOfCheck.h"
+#include "UseNamedCastCheck.h"
 
 namespace clang::tidy {
 namespace readability {
@@ -148,6 +149,8 @@ public:
         "readability-uppercase-literal-suffix");
     CheckFactories.registerCheck<UseAnyOfAllOfCheck>(
         "readability-use-anyofallof");
+    CheckFactories.registerCheck<UseNamedCastCheck>(
+        "readability-use-named-cast");
   }
 };